refactor: extract command execution into internal/execx
First cut of the package split: the runner stack (CommandRunner,
StreamingRunner, ExecRunner) moves to internal/execx, the one part of
pkg/agent with no dependency on agent types. Type aliases in pkg/agent
keep every existing reference working, so this establishes the
internal/ structure without churning consumers; follow-up extractions
peel off further leaves the same way.

BuildAgentEnv stays in pkg/agent: it is agent-environment policy over
Config, not exec mechanics.

The extracted package gains direct unit tests - process env overlay,
stdin, line streaming, the >64KB long-line scanner growth path, and
the stderr-only-on-failure contract - which previously only had
indirect agent-level coverage.
